Move more internal components into `util` folder (#33)
Assuming the top-level `include/rbs` holds public components (e.g.
`parser.h`), and `include/rbs/util` holds internal components, more
stuff should be moved under `util`.

This follows the same convention `prism` uses.

### Before

```
include
├── rbs
│   ├── ast.h
│   ├── defines.h
│   ├── encoding.h
│   ├── lexer.h
│   ├── parser.h
│   ├── rbs_buffer.h
│   ├── rbs_encoding.h
│   ├── rbs_location.h
│   ├── rbs_location_internals.h
│   ├── rbs_string.h
│   ├── rbs_strncasecmp.h
│   ├── rbs_unescape.h
│   └── util
│       ├── rbs_allocator.h
│       ├── rbs_assert.h
│       └── rbs_constant_pool.h
└── rbs.h
```

### After

```
include
├── rbs
│   ├── ast.h
│   ├── defines.h
│   ├── lexer.h
│   ├── parser.h
│   ├── rbs_location.h
│   ├── rbs_string.h
│   └── util
│       ├── rbs_allocator.h
│       ├── rbs_assert.h
│       ├── rbs_buffer.h
│       ├── rbs_constant_pool.h
│       ├── rbs_encoding.h
│       └── rbs_unescape.h
└── rbs.h
```
