+++ ext/rbs_extension/ast_translation.h
@@ -12,7 +12,6 @@
 #include "ruby/encoding.h"
 
 #include "rbs/ast.h"
-#include "rbs/rbs_encoding.h"
 #include "rbs/rbs_location.h"
 
 /// A bag of values needed when copying RBS C structs into Ruby objects.
