+++ include/rbs/util/rbs_encoding.h
@@ -6,8 +6,6 @@
 #ifndef RBS_RBS_ENCODING_H
 #define RBS_RBS_ENCODING_H
 
-#include "rbs_strncasecmp.h"
-
 #include <assert.h>
 #include <stdbool.h>
 #include <stddef.h>
