+++ src/util/rbs_encoding.c
@@ -1,5 +1,7 @@
-#include "rbs/rbs_encoding.h"
+#include "rbs/util/rbs_encoding.h"
 #include "rbs/util/rbs_assert.h"
+
+#include <ctype.h>
 
 #if defined(__GNUC__)
 #   define RBS_ATTRIBUTE_UNUSED __attribute__((unused))
@@ -5024,6 +5026,35 @@ const rbs_encoding_t rbs_encodings[] = {
     }
 #endif
 };
+
+/**
+ * Compare two strings, ignoring case, up to the given length. Returns 0 if the
+ * strings are equal, a negative number if string1 is less than string2, or a
+ * positive number if string1 is greater than string2.
+ *
+ * Note that this is effectively our own implementation of strncasecmp, but it's
+ * not available on all of the platforms we want to support so we're rolling it
+ * here.
+ *
+ * @param string1 The first string to compare.
+ * @param string2 The second string to compare
+ * @param length The maximum number of characters to compare.
+ * @return 0 if the strings are equal, a negative number if string1 is less than
+ *     string2, or a positive number if string1 is greater than string2.
+ */
+static int
+rbs_strncasecmp(const uint8_t *string1, const uint8_t *string2, size_t length) {
+    size_t offset = 0;
+    int difference = 0;
+
+    while (offset < length && string1[offset] != '\0') {
+        if (string2[offset] == '\0') return string1[offset];
+        if ((difference = tolower(string1[offset]) - tolower(string2[offset])) != 0) return difference;
+        offset++;
+    }
+
+    return difference;
+}
 
 /**
  * Parse the given name of an encoding and return a pointer to the corresponding
