+++ templates/ext/rbs_extension/ast_translation.h.erb
@@ -5,7 +5,6 @@
 #include "ruby/encoding.h"
 
 #include "rbs/ast.h"
-#include "rbs/rbs_encoding.h"
 #include "rbs/rbs_location.h"
 
 /// A bag of values needed when copying RBS C structs into Ruby objects.
